requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
decorator==5.1.1
frozenlist==1.4.1
fsspec==2024.2.0
gcsfs==2024.2.0
google-api-core==2.17.1
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.15.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
h3==3.7.6
idna==3.6
joblib==1.3.2
multidict==6.0.5
numpy==1.26.4
oauthlib==3.2.2
pandas==2.2.1
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
threadpoolctl==3.3.0
tzdata==2024.1
urllib3==2.2.1
yarl==1.9.4
